# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUser:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
  models.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Message:
    properties:
      _id:
        type: string
      content:
        type: string
      recever:
        type: string
      sender:
        type: string
    type: object
...
  models.SendMessageM:
    properties:
      content:
        minLength: 5
        type: string
      recever:
        type: string
      sender:
        type: string
    required:
    - content
    - recever
    - sender
    type: object
...
  models.UserModel:
    properties:
      _id:
        type: string
      bio:
        type: string
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      following:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
      password:
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
host: localhost:5000
info:
  contact: {}
  description: This is Swagger docs for rest api golang fiber
  title: Fiber Golang Mongo Grpc Websocet etc..
  version: "1.0"
paths:
  /chat/get-user-unreadedmsg:
    get:
      consumes:
      - application/json
      description: Get unreaded message count & recodes for user
      parameters:
      - description: user id
        in: query
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get unreaded message count & recodes for user
      tags:
      - Chat
  /chat/getmsgsbynums:
    get:
      consumes:
      - application/json
      description: GetMsgsByNumbetween two users by pagenation
      parameters:
      - description: Staring point page num
        in: query
        name: from
        required: true
        type: integer
      - description: first user id
        in: query
        name: firstuid
        required: true
        type: string
      - description: second user id
        in: query
        name: seconduid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: get message by pagenation
      tags:
      - Chat
  /chat/mark-msg-asreaded:
    get:
      consumes:
      - application/json
      description: mark messages as read for user uupate the recoded make is read
        true num 0
      parameters:
      - description: main user id
        in: query
        name: mainuid
        required: true
        type: string
      - description: ohter user id
        in: query
        name: otheruid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: mark messages as read for user
      tags:
      - Chat
  /chat/sendmessage:
    post:
      consumes:
      - application/json
      description: SendMessage form one user to another
      parameters:
      - description: user SendMessage deatils
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.SendMessageM'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: send message to friend user
      tags:
      - Chat
...
  /user/signin:
    post:
      consumes:
      - application/json
      description: Login an user by providing email, password
      parameters:
      - description: user Login deatils
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: login a  user
      tags:
      - Authentication
  /user/signup:
    post:
      consumes:
      - application/json
      description: Register an ew user by providing email, password , first name ,
        last name
      parameters:
      - description: user register deatils
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Gegister a new user
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"